requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=9.5.0
numpy>=1.21.0
opencv-python>=4.8.0
tqdm>=4.65.0
concurrent-futures>=3.1.1
accelerate>=0.20.0
qwen-vl-utils>=0.0.1
